ChainOfAssignments.java: aligned changes to the VertexName constructor.
ReachingDefinitionsHandler.java: Implemented a new algorithm (graph-based) to produce the data that is to be displayed in the interface. Also new filter have been applied in order to reduce the no. of redundant messages in the interface. Classes ConfigurationEdgeFactory.java, ConfigurationEdgeNameProvider.java and ValueContainerEdge.java are part of this implementation. It still needs to be tested. Too tired now :|

git-svn-id: https://subversion.assembla.com/svn/cide_ei@9 d7039242-6a02-407b-9b27-e99ad9532730